// typescript-poo/src/app/models/prodcut.model.ts
import { Category } from "./category.model";

export interface Product {
    id:          number;
    title:       string;
    slug:        string;
    price:       number;
    description: string;
    category:    Category;
    images:      string[];
    creationAt:  Date;
    updatedAt:   Date;
}

// https://app.quicktype.io/